fixup: Remove ServerBump::request to simplify
The field description claimed it is required by the Client API. AFAICT,
that description was misleading because ServerBump::request is not fed
directly to Client. The original text said "server-side", not Client.
Back then, the Ssl::ServerPeeker::request field was used for
storeCreateEntry() and FwdState::fwdStart() calls. Today, similar
contexts have access to ClientHttpRequest and its HttpRequest.