Invalidate all cached variants using STORE_META_VARY_ID
When a parent vary object is created, it acquires a UUID, which is
subsequently copied to all its child objects. When the parent object
is purged and created again, all its child objects become invalid due to
the UUID mismatch. Before this fix, the parent vary object recreation
did not invalidate its children.