Fix client_side_request.cc:1683: "request != newRequest" assertion
Adaptation code normally clones the virgin message even if no adaptation
is necessary. ClientHttpRequest::handleAdaptedHeader() relied on that
approach. However, there are cases where adaptation routing code cannot
forward the message to the next routing service and leaves the message
object intact.

This "left the message object intact" outcome is possible (at least!)
when the routing plan is set by the adaptation service itself via the
X-Next-Services meta-header, and the "imported" plan includes adaptation
services that are not applicable to the current vectoring point. See
thePlan.exhausted() branch of Adaptation::Iterator::step().

We possibly could change the adaptation routing code to prevent such
outcomes, but since "leave the message intact" adaptation decisions are
very common, it is better to optimize their handling instead (by
returning the virgin message object to signal no-op adaptation).

With this change, that optimization is implemented for eCAP. ICAP code
should be optimized as well, but that far more complex change is left as
a TODO. The receiving code now does not assume that the virgin and
adapted message objects are always different.

XXX: It is not clear whether Client::handleAdaptedHeader() is ready for
this change.