Use anonymous Union in Optional
The named Union introduced in b2cd19a had a problem.  Since it was
stored inside Optional as a separate object, it could not destroy its
value_ member depending on the Optional::hasValue_ flag. That
implementation was wrong because ~OptionalStorage() always destroyed
value_, even if it had not been initialized.

Anonymous Union, on the other hand, injects its members into the
enclosing class (Optional in our case), making it possible to correctly
destruct the value_ member in ~Optional() itself.