Fixed a branch bug where HttpReply was unexpectedly destroyed
This happened because the new code in HttpReply::parseHeader()
passed the reply to ACLFilledChecklist. Despite the locking,
the list unlocks (and deletes) the reply in destructor.

The underlying problem looks more general -
we should probably refactor HttpStateData::processReplyHeader(),
where a raw HttpReply is created but not owned until, e.g.,
setVirginReply() takes the ownership, thus allowing to delete
the reply accidentally.  I added a TODO about this.