fixup: Simplified memReaderHasLowerOffset() replacement logic
Earlier in the branch, we replaced that with memHeaderOffsetLowerThan()
to hide external access to copyInto.offset (which stores info that is
passed in async calls), but I found a simpler way to achieve that.

Technically, store_client::getType() is public so LowestMemReader could
use that directly instead of adding x.reliesOnReadingFromMemory(), but
the true meaning of getType() result is so complicated that I decided it
is best to avoid using that in an external code.

Also fixed branch-broken "make check".