Simplified code and reduced diff
store_client orchestrates cooperation of three asynchronous players:

* Sink: The Store client (store_client owner)
  STCB callback is requested via a storeClientCopy()/copy() call.
  Set _callback.callback_handler implies waiting for this callback.

* SourceA: Store disk subsystem
  storeClientRead*() callbacks are requested via a storeRead() call.
  flags.disk_io_pending implies waiting for these callbacks.

* SourceB: Store memory subsystem:
  Copying is "spontaneously" triggered by StoreEntry::invokeHandlers().
  This copying is synchronous. No flag guards its progress.

SourceA and SourceB copying are mutually exclusive (KISS for now).

We can finalize asserts while avoiding a lot of code modifications and
complications if we limit each copy() request to a single answer,
without updates, even though multiple invokeHandlers() calls are
possible while we are waiting for the new notifier to be called.
Eventually, such updates may be supported, but it will probably require
changing STCB callback API to make it possible to answer with
data-and-error or data-and-EOF combination. Such STCB changes will
propagate to callers and deserve a dedicated PR.

This no-updates approach means that the creation of the notifier call
effectively ends processing -- store_client just waits for that call to
fire so that it can relay the answer to synchronous STCB. The notifier
is now restricted to being internal notification implementation detail.
When STCB itself becomes asynchronous, this logic will continue to work.

Also removed callback_data from the Callback::pending() condition. It is
conceptually wrong to require non-nil callback parameter, and it is
never cleared separately from the callback_handler data member anyway.