fixup: Diff reduction
This change also restores the (meaningfulness of) the (anEntry == entry)
assert in doCopy(). The restored assert in storeClientCopy2() is
excessive because doCopy() now asserts the same thing, but I think it is
better to have two identical asserts here than to imply that we are
relaxing storeClientCopy2() preconditions. Ideally, storeClientCopy2()
itself should be removed.