fixup: Reduced diff by restoring store_client::cmp_offset
The field was removed in an earlier branch commit because it was
considered a "useless debugging code for the merged branch" that was
merged long time ago (528b2c6). I believe that field was meant to be
used for self-checking Squid offset management code (during and after
that branch conversions), but those checks could not have been enabled
by default because they do not work for Range requests that may "skip"
portions of the continuous response body.

I think the field should be removed, but others may might still consider
it useful for validation (in Squids built with ONLYCONTIGUOUSREQUESTS).
Since removing it is out of scope, and the field does not complicate our
work much, I felt we cannot easily justify its removal in _this_ branch.