Revert previous commit "Make store_client destruction async"
The intent of that change was correct, but storeUnregister() is not the
only entry synchronous entry point into store_client. Instead of writing
quite a bit of custom code to cover every point, we should convert
store_client into an AsyncJob. That conversion will require some API
adjustments. It deserves a dedicated PR.