Restored store_client::flags.store_copying
The flag was removed earlier in the branch because we thought that
making callbacks asynchronous is sufficient to eliminate the need for
it, but the flag is also necessary to prevent invokeHandlers() from
re-entering store_client while it is still inside doCopy().

Checking this flag in invokeHandlers() is still not enough to guarantee
that we will not reenter the same store_client. For example, if we start
calling StoreEntry::write() instead of MemObject::write(), we may.