Print URI scheme when detailing port configuration context
In ef6fceb we removed port information from several level-1 lines, to
avoid duplication with code context details. Since URI scheme was a part
of the old debugging, we need to print it with the context now.

Old PortCfg::name stored either the port address (by default) or the
internal port name (if configured). Since we need to prefix only address
with the corresponding URI scheme, we had to split the original
PortCfg::name into two fields, for 'name' and 'specification'
accordingly.