Do not leak caller context to other transactions in helperHandleRead()
When the transaction A starts a new helper instance, it schedules
helperHandleRead() read callback and waits for the helper reply.  After
receiving reply from the helper, it re-schedules the call again with A
context, thus propagating it to the next and all future transactions,
reading from this helper.

So we need to:

* Run the general helper initialization code, such as starting
  servers and I/O scheduling without any specific context.

* Store the context before requesting and restore it after receiving
  response from the helper.