Removed PortIterator
We faced several complications/difficulties when trying to implement
context-switching port iterators. For example, their scope was limited
to for-range loops, the implementation required to make CodeContextGuard
movable, etc. So, for simplicity sake, we decided to abandon the
over-complicated and error-prone iterator-based approach and manage
context in each loop manually, by means of CodeContextGuard object.