Removed CodeContextGuard
Unfortunately, the solution with CodeContextGuard in loops introduced in
30d271f has a couple of flaws:

* The context is switched back and forth at each iteration.

* The throwing context becomes lost because CodeContextGuard
  automatically restores it to the original one during stack unwinding.

To fix this, we need to manage code context manually in each loop
(introducing code duplication). We can probably avoid this duplication
by converting intrusive HTTP/FTP port lists into containers and
refactoring the loops accordingly. For now, we decided to get by the
simpler approach.