fixup: Restore calling updateAnchoredWith() for MemStore
... until we can get rid of the copyFromShm() preloading hack.

    assertion failed: store_client.cc:227:
    "entry->hasDisk() && !entry->swapoutFailed()"

We still need to preload the freshly anchored entry for the Store to get
what it expects for memory-cached entries. There might be bugs in that
preloading logic, but not the bugs we are fixing now.

The rock store is different because it does not have the preloading hack
Its entry.swap_file_sz update is already covered by anchor.exportInto()
in Rock::SwapDir::anchorEntry() that SwapDir::anchorToCache() calls.