fixup: Do not anchor the same entry to two stores
The primary branch fix missed a case where Controller::anchorToCache()
is asked to anchor an entry that is already anchored to a disk cache. We
correctly excluded each store from double-anchoring, but still allowed a
memory cache to anchor an entry already anchored to a disk. The latter
is not yet supported because different stores might contain different
metadata, and the code is probably not ready to deal with such metadata
conflicts.

To fix that, we have to avoid anchoring already anchored entries at
Controller level, just like Controller::syncCollapsed() already does.
That means that Controlled::anchorToCache() can stop accommodating
repeated anchoring attempts. TODO: I am not sure the Controlled change
part is desirable; it increases risk and will probably require a
reversal later, when Squid starts supporting anchoring to multiple
stores.