fixup: Minor polishing
"memCache.io == MemObject::ioDone" does not mean that the entire object
is currently in the local transient memory. It only means that it was
fully loaded there at some point in the past. It may be gone now. It may
not even be in the shared memory cache anymore. The difference is not
important for this code -- we do not want to anchor the same entry
multiple times anyway (at least for now, for simplicity sake) -- but
debugs() should not mislead the reader.