Refactored to reduce locked entry time and simplify/avoid complications
It is too dangerous (and probably just flat out wrong) to stuff a
StoreEntry pointer into std::unique_ptr (that is supposed to auto-unlock
the entry via Deleter) because std::unique_ptr does not call the
unlocking Deleter when release()-ing the pointer, losing one lock unless
the release() caller is very careful. That smart pointer design is just
not compatible with reference-counting objects, some would say by
["unique"'s] definition.

We do need a RefCount-like smart pointer for StoreEntry pointers, but
introducing one is difficult and this branch is not about that.