Revert "Build an Ssl::ServerBump object on all of the bumping cases"
This reverts branch commit c3bd63a4d1d9f9bc95e280f9e9dc58598affeb15. The
intent behind that commit was absolutely correct. I have not checked its
implementation carefully, but I think such a change is too risky for
this v5 backport -- there may be code relying on sslServerBump existence
as a sign that we are not in client-first or server-first bumping mode.
The final implementation will also need to rename that field, which will
cause widespread non-changes.

TODO: Provide a replacement for the missing sslServerBump object.