A storage for vary-related attributes
Attributes such as vary-mark and uuid should be kept in sync.  The
synchronization depends on the entry type, for example, 'base' entry
goes without vary-mark, while leaf entries require it. For this purpose,
a new VaryDetails class was created, allowing to construct two
types of vary attributes (for leaf and base entries respectively).