fixup: Avoid ErrorDetails() construction ambiguity/danger
The ErrorDetails copy constructor is safe and efficient. It is also
default-generated (for now; and we can default it explicitly if
RefCount/Lock become Interfaces, as they should).

The ErrorDetails(const ErrorDetail &) conversion constructor worked
fine, but it was too dangerous because somebody could call it by
dereferencing a ErrorDetailPointer while thinking that they are calling
the ErrorDetails copy constructor. In other words, somebody could easily
create an ErrorDetails object that contains another ErrorDetails object
(instead of "flattening" the latter like a copy constructor does). We
now prevent such errors at compile time.