fixup: Do not end ended request satisfaction
    assertion failed: BodyPipe.cc:121: "p != NULL"
    in xassert() at debug.cc:624
    in BodyConsumer::stopConsumingFrom() at BodyPipe.cc:121
    in ClientHttpRequest::endRequestSatisfaction() at client_side_request.cc:2234
    in ClientHttpRequest::storeEntry() at client_side_request.cc:1700
    in clientReplyContext::removeClientStoreReference() at client_side_reply.cc:197

In the previous branch commit, I did not fully realize that
endRequestSatisfaction() itself does not end our entry_ usage. It only
ends receiving of the response body (and the active subscription to
receive that body is the call precondition). We may continue to use the
entry_ after endRequestSatisfaction() (e.g., for sending the response
body to the client) and that use does not mean we should call
endRequestSatisfaction() again when switching the entries.

Also tightened assertions/checks after realizing that all the
endRequestSatisfaction() callers in the official code are dealing with
request satisfaction mode. Regular request adaptation is handled in
Client -- Server classes do not setConsumerIfNotLate() except in request
satisfaction cases.

Also explicitly handle "completed" entries in endRequestSatisfaction().
The earlier code worked, but "completing" a "completed" entry creates
quite a bit of lengthWentBad() noise. This early check really belongs to
the StoreEntry::complete*() family of functions, but that code probably
needs to be refactored first (e.g., to avoid checking twice).

Also added the check for aborted request satisfaction to
ClientHttpRequest::resumeBodyStorage(), missed in the previous branch
commit.