Fix milliseconds in early saved level-2+ cache.log messages
The milliseconds part of an early message time was wrong because
debugLogTime() got it from the current_time global instead of the saved
message timestamp. To fix that, we adjusted
DebugMessageHeader::timestamp to store time with microsecond precision.

The problem could manifest itself only for early saved messages, when
there may be a noticeable time gap between message creation and message
writing.