Close idle persistent connections to a CachePeer after reconfigure
The Connection object representing connection to a cache_peer has a
cbdata-protected CachePeer, becoming invalid after reconfiguration (when
old CachePeer objects are removed). Unaware of this,
HttpStateData sets its toOrigin flag and communicates to the peer
(via the old persistent connection) as with origin. This, in fact,
leads to the request URI built in the origin-form (instead of
absolute-form) and subsequently 400 Bad Request response
from the peer.

Now we close all persistent connections associated with removed
CachePeers.