Do not kill Squid if parameters() file is empty or missing
We can provide ACL-specific arguments from external files in two ways:

* with the 'parameters' option:
    acl allowlist dstdomain parameters("/path/to/filename")
* with the file name in double quotes:
    acl allowlist dstdomain "/path/to/filename"

If the filename is empty or missing, ConfigParser aborts Squid
with a configuration error in the first case but just shows a
level 0 error message in the second. We need to handle both
cases similarly and let the ConfigParser caller decide whether
an empty or missing configuration is fatal or not.