Moved ConfigParser::SetAclKey() to Acl namespace
and moved the 'key extracting' semantics to each caller.

The method initially was designed to handle required ACL attributes.
However, we need also to cover optional ACL attributes, such as in
ACLCertificateData. In such cases, we cannot determine whether the
extracted (optional) token is a genuine ACL key (and thus subject to
SetAclKey()) without extracting it first.