Optimize and avoid WARNINGs when configuring large acl regexes
    WARNING: optimisation of regular expressions failed; using fallback
    method without optimisation

When adding the last regex on the acl line made the total accumulated
regex size larger than 1024 characters, Squid logged the above warning
and refused to "optimize" the acl line. The simplest reproducer is an
acl line with a single regex longer than 1024 characters.

----

This fix is based on master/v6 commit 0fa036e that contains a lot more
(and more invasive) improvements.