Remove RemovalPolicy::Dereferenced()
This method was used by Store::Storage subclasses to signal
the Policy that the StoreEntry became idle. Actually, in both lru and
heap policies it it was implemented as RemovalPolicy::Referenced(),
merely updating/refreshing the element order in the Policy storage.

However, with heap policy, calling Referenced() when the entry becomes
idle looks useless because its position in the underlying tree storage
would not change. With lru policy, calling this method when the entry
becomes idle also looks unnecessary because we should have called it
while acquiring the entry from Store.