Documentation: Polish "refresh_pattern percent" description
The original text contained a "last modification age" typo and was
misinterpreted by some admins as if Squid applied the configured percent
to the current object age.

The text also did not make it clear that the percent-based heuristic is
effectively only applied to responses with a Last-Modified header (in
addition to the usual "without an explicit expiry time" precondition).

----

Cherry-picked v5 commit 0d077980672c0433f31ecf7c6e849dd1043c2281.