Run cache_peer probes without transaction context
Even when triggered by a master transaction, these background probes
should not be associated with it. That transaction just wants to know
whether the neighbor is "up" right now. The transaction does not wait
(and is not aware of) the background probe its question may or may not
have triggered.

----

Cherry-picked master/v6 commit be5328880c80c64300221f1919b8cb78981d3b78.