Primary fix for %err_code missing in ERR_RELAY_REMOTE transactions
Earlier related branch fixes do not hurt, and may be fixing other
(unknown) problems or (at least) moving that code in the right
direction, but the "%err_code missing in ERR_RELAY_REMOTE transactions"
bug itself was caused by the BuildHttpReply() not updating
HttpRequest::error on ERR_RELAY_REMOTE transactions (because
pre-computed response in those transactions triggered a premature exit
from BuildHttpReply()).

BuildHttpReply() should not be responsible for detailing errors at all
because its task is building the error response, but the problem is even
bigger -- ErrorState class as a whole should not be responsible for
updating HttpRequest and ALE. Despite its name, the objects of this
class may not even be created for certain errors. Fixing that requires a
lot more changes, so we are just leaving an XXX for now.