CI: Remove unnecessary test-functionality test wrappers
These workarounds are not needed for the current and future code in this
branch. Other branches get their own test-functionality.sh files that
can be used to maintain a branch-specific collection of test wrappers.

The (now unused) has_commit_by_message() function was left in the script
because that function is likely to be used by future workarounds. Unlike
specific test workarounds that only apply to a subset of old code, this
and similar functions can be viewed as a reusable code "library".

----

Cherry-picked master/v7 commit 15cf09067d1ede6b482368b9665e304102c15783
and resolved a conflict by removing check_cache_refresh_response() that
was specific to (and never needed for) this branch (see commit bf2aecc).
That wrapper was added to this branch when we were still adding such
unnecessary wrappers for _other_ branches to reuse.