fixup: Avoid adding StoreEntry::appendList()
StoreEntry should not be dealing with content formatting. We do not want
to maintain dozens of appendFoo() methods, for all sorts of information
we need to format and add to store!

AFAICT, this change also fixes extra trailing spaces when official code
dumps "acl" lines for ACLs that have no options and no parameters (e.g.,
connections_encrypted ACL).