fixup: Undo dump_acl:... debugging changes
Besides ACLs, only outdated WCCP directives have this (arguably
excessive) debugging AFAICT. This debugging should probably be removed
completely, but that is out of scope here.

For now, we should just keep it at its original location inside
dump_acl(), where it matches the function name and applies to built-in
ACLs as well.