Added (negative) configuration test cases
Also polished the related error text after seeing how the two applicable
use cases are rendered. I am not fully satisfied  with the current
version, but it is difficult to cover two different use cases with the
same error message. We will come back to this if others request changes.