fixup: Acl::CacheManagerCheck simply has no parameters to dump()
Calling that (officially misnamed) dump() method is not really a bug. In
fact, one could imagine some future scenarios where the manager ACL gets
some (hard-coded) parameters that could be reported. That reporting
could not happen in regular squid.conf dumping context, but the dump()
method does not have to be specific to that context.

See also: Acl::ConnectionsEncrypted.