Removed "manager_type" ACL type
The manager ACL has no type that can be expressed in squid.conf syntax.

I have considered alternative B: Provide a type name that can be used in
squid.conf (e.g., "manager_acltype") and, naturally, allow folks to
define new ACLs with that ACL type (e.g., `acl mgr manager_acltype`).

That alternative B is more complex than the current "built-in type"
because

* Alternative B requires writiong parsing code for those ACLs.

* Alternative B may require writing code to dump() the manager ACL
  (because that ACL will gain a squid.conf-compatible syntax!).

* Alternative B requires dealing with multiple definitions of same-name
  built-in ACLs.

* Alternative B might complicate or even preclude some future
  optimizations for other built-in ACLs: For some built-in types, it
  might be useful to assert that there is only _one_ registered ACL
  object of that type.

We might find a reason to move towards that alternative in the future,
but, for now, I do not know of any. And such a move would not require
changing existing configurations.

N.B. It is possible to "alias" built-in ACLs now, without creating more
built-in ACL object: `acl mgr all-of manager`. We are _not_ restricting
admin ability to introduce new ACLs with admin-friendly names that match
just like the "manager" ACL.