fixup: Remove ConnStateData::sslBumpMode to simplify
Now-available ServerBump already represents this state.

ConnStateData::sslBumpMode removal required adjusting ServerBump to
store the "current" SslBump action because we still need to know what
the requested action is when we cross the step boundary (and the
now-current step action becomes not-yet-set/unknown).

TODO: We may be able to get the same info from the requested_ actions if
we interpret them specially based on the caller needs. I also have not
checked yet whether always ignoring not-yet-set steps is going to work
now that the Squid-server part of server-first processing is in step1.

TODO: If we end up not reusing class Actions, we should not define it as
if it is being used for two (naturally different) fields.

----

Cherry-pick bag62 commit a032559.