fixup: SslBump steals doCallouts errors _after_ bumping
... resulting in Squid effectively ignoring REQMOD-rejection of
successfully _bumped_ GETs in a client-first.none.none test.

In doCallouts() and probably some other places, the serverBump presence
is mistaken for a "we are on the SslBump path" condition. That logic
works only before and during SslBump application, not after. Once the
connection is bumped (or spliced, but there are fewer opportunities for
error there), the presence of "done" serverBump is just a historical
artifact for logging and such, not a signal to continue SslBump
processing!

The comprehensive fix for this bug will look different, but we can
reduce the bug surface and code complexity by moving SslBump-path error
storing to (somewhat misnamed) sslBumpAccessCheck() which is more
careful about checking whether SslBump processing is done.

Also supply log_uri (not two store_ids) when storing doCallout errors,
as prescribed by storeCreateEntry() API. I do not know what exact bugs
this change is fixing.

----

Cherry-picked bag62 commit 70f2790 and resolved simple conflicts.