Drop redundant LIBADD_DL overwrite
The removed code questions whether it is needed.

The answer is no. LIBADD_DL is already set by one of the LT_INIT
macros to the same value we are overwriting it to.

Except, in the case where the user has supplied a custom libdl
location this code broke their build.

----

Cherry-picked master/v6 63a76d3f68cd904951dc07d230b84f3047b01521.