Maintenance: code style updates in libcompat
The libcompat code is targeted here specifically because it is
rarely touched so is not likely to be soon updated by our
"upgrade when modified" policy for source style changes.

Doing this part of the code semi-manually also allows us
to trivially exclude the directory from automated conversion
attempts and not bother with automating the detection of
whether a .h is included in C files.

----

Cherry-picked master/v6 commit f70cf39b14872dddac2e14c2858bded78429150a
and resolved low-complexity conflicts.