Add slow ACL support to on_unsupported_protocol directive
The primary use case is outsourcing tunnel-vs-terminate decisions to
external ACLs that may, for example, analyze %>handshake bytes.

No functionality (or performance) changes for Squid configurations
without on_unsupported_protocol rules.

This change might also correct on_unsupported_protocol behavior when
pipeline_prefetch is on. Tunneling code assumes that it is dealing with
the only parsed request on the client-to-Squid connection. We are not
certain whether other on_unsupported_protocol tunneling preconditions
made that hidden assumption true. The adjusted code makes it explicit.