Delay SslBump to authenticate CONNECTs
In branch commit 20eb753e, we ensured that "ConnStateData::sslServerBump
is always set for transactions that are subject to SslBump processing",
but we went too far, calling startSslBumpProcessing() too early (in
clientProcessRequest): If a CONNECT request needs to be authenticated,
then it is not "subject to SslBump processing".

That mistake triggered ClientRequestContext::sslBumpAccessCheck(). That
method had to call ServerBump::noteFinish() to prevent SslBump
processing for "authenticating CONNECT". That, in turn, forced
sslBumpAfterCallouts() to treat tlsBumpDone specially to prevent calling
inapplicable sslBumpAfterCalloutsAtStep*() that would assert. That
special treatment was buggy: "Authenticating CONNECT" transactions got
stuck because they were not reading from the client.

Now we delay sslServerBump creation until we receive an authenticated
CONNECT request. sslBumpAfterCallouts() still treats tlsBumpDone
specially, but now that is done to handle nested CONNECTs (i.e. CONNECTs
that are received on a successfully bumped client connection). That
handling was a TODO. It still needs more work.