Undo "create adaptation history object if adaptation_meta is configured"
... change in branch commit 0f0ddd98.

There are problems with that method, but I do not think we have to solve
any of them on this branch (and, hence, we should not). Instead, we
should use adaptHistory(true) instead of adaptLogHistory() where it
matters for solving in-scope/branch problems. The former method not only
works better, but it is actually correct (or at least more appropriate)
for setting annotations than %code-focused adaptLogHistory()!