Revert branch adaptation_meta changes
Annotation handling in Squid is too inconsistent to improve that
directive in this branch. Our attempt ran into too many questions and
XXXs (see previous branch commit fa578739 for an incomplete list). It is
best to keep that directive unchanged in this branch.

The initial branch adaptation_meta changes were driven by code
duplication avoidance, but we are not going to reuse adaptation_meta
code for handling adaptation-sent annotations. We now plan to reuse code
written to handle helper-sent annotations (because ICAP/eCAP transaction
are a lot more like helper transactions than they are like
adaptation_meta evaluations)...

XXX: This commit breaks build because it removes in-use
Adaptation::History::updateMetaHeader().