annotate_client did not annotate error:* transactions
cc71dd7 introduced an admin-level BUG requirement that a non-empty
message pipeline means an existing HttpRequest object. That requirement
was incorrect because when the pipeline receives the HttpStream object
(returned by ConnStateData::parseOneRequest()), the corresponding
HttpRequest is not created yet (an may be not created at all, depending
on further checks). Now, instead of demanding HttpRequest
presence in annotate_client, we annotate requests if/when they get
created (and imported into ClientHttpRequest) by copying the existing
connection annotation. That copying already worked for successfully
parsed and fake HttpRequests, but was missing for error-uri requests
(such as "error:invalid-request").

Also fill these 'error:*' HttpRequests with connection-level
information. For example, access_log takes this information from
the request, and fails to match some ACLs, such as 'src'.